+++ src/Main.hs
@@ -56,7 +56,7 @@ health = responseLBS
     (encode okHealthResult)
 
 notFoundPage :: Network.Wai.Response
-notFoundPage = responseLBS status404 [(hContentType, "text/html")] "<h1>Page not found</p>"
+notFoundPage = responseLBS status404 [(hContentType, "text/html")] "<h1>Page not found</h1>"
 
 data HealthStatus = Ok | Warning | Critical