+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # Revision history for meetup-proxy
 
+## NEXT   -- UNRELEASED
+
+* Fix markup for 404 page
+
 ## 1.0.0  -- 2017-11-06
 
 * First version. Released on an unsuspecting world.